Revert "change loot type"
This reverts commit 417bb6c9e168f19e58f1e0dbaf0e9e92a21347c5.

+++ sprint3/problems/find_return/solution/src/model/loot_type.h
@@ -15,7 +15,7 @@ struct LootType {
   std::optional<int> rotation;       // example -> = 90;
   std::optional<std::string> color;  // example -> = "#338844";
   std::optional<double> scale;       // example -> = 0.03;
-  std::optional<size_t> value;
+  std::optional<int> value;
 };
 
 }  // namespace model
